pyproject.toml
[project]
name = "elevenlabs-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "elevenlabs>=2.23.0",
    "fastapi[standard]>=0.121.2",
    "groq>=0.35.0",
    "python-dotenv>=1.2.1",
]
